requirements.txt
Flask==2.3.3
Authlib==1.2.0
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
psycopg2==2.9.10
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
pytz==2023.3
requests==2.32.5